requirement.txt
python-telegram-bot==21.6
Flask
gunicorn
python-dotenv

